--merge command-line argument now functional (but needs testing!)
With the --merge argument, any TrueType programming in the input font (including fpgm, prep, cvtg, cvar and all glyph programs) is retained, and Xgridfit programming is layered on top of it. This enables you to autohint a large font (say, with VTT or ttfautohint) and supply higher-quality Xgridfit hints to a selection of glyphs without disturbing the autohinting.